Add order inside of the code and generate new group for actions - separated from animations
