+++ godtools/Views/TractElements/TractCardsAnimations.swift
@@ -11,80 +11,7 @@ import UIKit
 
 extension TractCards {
     
-    // MARK: - Animations for the cards inside of the Cards container
-    
-    func setEnvironmentForDisplayingCard(_ card: TractCard) {
-        changeToOpenCards()
-        
-        var foundCard = false
-        
-        for element in elements! {
-            let elementCard = element as! TractCard
-            if card != elementCard {
-                if foundCard {
-                    elementCard.hideCard()
-                } else {
-                    elementCard.showCard()
-                }
-            } else {
-                foundCard = true
-            }
-        }
-        
-        if card == elements?.last {
-            showCallToAction()
-        } else {
-            hideCallToAction()
-        }
-    }
-    
-    func resetEnvironment() {
-        changeToPreviewCards()
-        
-        for element in elements! {
-            let elementCard = element as! TractCard
-            elementCard.resetCard()
-        }
-    }
-    
-    func showFollowingCardToCard(_ card: TractCard) {
-        var foundCard = false
-        
-        for element in elements! {
-            let elementCard = element as! TractCard
-            if card == elementCard {
-                foundCard = true
-                continue
-            }
-            if foundCard {
-                elementCard.showCard()
-                break
-            }
-        }
-    }
-    
-    func showCallToAction() {
-        let rootView = self.parent as! TractRoot
-        rootView.showCallToAction()
-    }
-    
-    func hideCallToAction() {
-        let rootView = self.parent as! TractRoot
-        rootView.hideCallToAction()
-    }
-    
-    // MARK: - Animations for the Cards container and the Header
-    
-    func changeToOpenCards() {
-        if self.cardsState == .open {
-            return
-        }
-        
-        self.cardsState = .open
-        
-        let rootView = self.parent as! TractRoot
-        rootView.hideHeader()
-        
+    func transformToOpenUpCardsAnimation() {
         UIView.animate(withDuration: 0.35,
                        delay: 0.0,
                        options: UIViewAnimationOptions.curveEaseInOut,
@@ -93,16 +20,7 @@ extension TractCards {
                        completion: nil )
     }
     
-    func changeToPreviewCards() {
-        if self.cardsState == .preview {
-            return
-        }
-        
-        self.cardsState = .preview
-        
-        let rootView = self.parent as! TractRoot
-        rootView.showHeader()
-        
+    func transformToInitialPositionAnimation() {
         UIView.animate(withDuration: 0.35,
                        delay: 0.0,
                        options: UIViewAnimationOptions.curveEaseInOut,
