+++ godtools/Views/TractElements/TractElementStaticFunctions.swift
