+++ godtools/Views/TractElements/TractModals.swift
@@ -10,12 +10,4 @@ import UIKit
 
 class TractModals: BaseTractElement {
 
-    /*
-    // Only override draw() if you perform custom drawing.
-    // An empty implementation adversely affects performance during animation.
-    override func draw(_ rect: CGRect) {
-        // Drawing code
-    }
-    */
-
 }
